requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# HTTP Client
httpx>=0.25.0

# Article Extraction
trafilatura>=1.6.0
newspaper3k>=0.2.8
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Browser Automation (optional, for JS-heavy sites)
playwright>=1.40.0

# Data Validation
pydantic>=2.5.0

# Utilities
python-dateutil>=2.8.0

# NLP for newspaper3k (optional)
nltk>=3.8.0
